src/utils/requirements.txt
google-ads>=20.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
serpapi>=0.1.0
google-cloud-language>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
playwright>=1.20.0
